chore: read addresses as 64-bit and fix API

+++ packages/tsprocess/lib/functions.cc
@@ -16,7 +16,7 @@ Napi::Value read_byte(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<int8_t>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read byte at %x", address)).ThrowAsJavaScriptException();
@@ -33,7 +33,7 @@ Napi::Value read_short(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<int16_t>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read short at %x", address)).ThrowAsJavaScriptException();
@@ -50,7 +50,7 @@ Napi::Value read_int(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<int32_t>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read int at %x", address)).ThrowAsJavaScriptException();
@@ -67,7 +67,7 @@ Napi::Value read_uint(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<uint32_t>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read uint at %x", address)).ThrowAsJavaScriptException();
@@ -84,7 +84,7 @@ Napi::Value read_float(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<float>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read float at %x", address)).ThrowAsJavaScriptException();
@@ -101,7 +101,7 @@ Napi::Value read_long(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<int64_t>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read long at %x", address)).ThrowAsJavaScriptException();
@@ -118,7 +118,7 @@ Napi::Value read_double(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto result = memory::read<double>(handle, address);
   if (!std::get<1>(result)) {
     Napi::TypeError::New(env, logger::format("Couldn't read double at %x", address)).ThrowAsJavaScriptException();
@@ -204,7 +204,7 @@ Napi::Value read_buffer(const Napi::CallbackInfo &args) {
   }
 
   auto handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
   auto size = args[2].As<Napi::Number>().Uint32Value();
   auto buffer = new uint8_t[size];
   auto data = (uint8_t *)malloc(sizeof(uint8_t) * size);
@@ -343,7 +343,7 @@ Napi::Value read_csharp_string(const Napi::CallbackInfo &args) {
   }
 
   void *handle = reinterpret_cast<void *>(args[0].As<Napi::Number>().Int64Value());
-  auto address = args[1].As<Napi::Number>().Uint32Value();
+  auto address = static_cast<intptr_t>(args[1].As<Napi::Number>().Int64Value());
 
   if (address == 0) {
     return Napi::String::New(env, "");
